backend/requirements.txt
# requirements.txt

uvicorn == 0.38.0
fastapi[standard] == 0.121.0
psycopg[binary] == 3.2.12
pydantic == 2.12.3
python-dotenv == 1.2.1
bcrypt == 5.0.0